End of section 5

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -56,12 +57,13 @@ public class AuctionsController : ControllerBase
     return _mapper.Map<AuctionDto>(auction);
   }
 
+  [Authorize]
   [HttpPost]
   public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
   {
     var auction = _mapper.Map<Auction>(auctionDto);
-    //TODO: add current user as seller
-    auction.Seller = "test";
+
+    auction.Seller = User.Identity.Name;
 
     _context.Auctions.Add(auction);
 
@@ -79,6 +81,7 @@ public class AuctionsController : ControllerBase
     return CreatedAtAction(nameof(GetAuctionById), new { auction.Id }, newAuction);
   }
 
+  [Authorize]
   [HttpPut("{id}")]
   public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
   {
@@ -91,7 +94,10 @@ public class AuctionsController : ControllerBase
       return NotFound();
     }
 
-    //TODO: check seller == username
+    if (auction.Seller != User.Identity.Name)
+    {
+      return Forbid();
+    }
 
     auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
     auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
@@ -100,7 +106,7 @@ public class AuctionsController : ControllerBase
     auction.Item.Year = updateAuctionDto.Year ?? auction.Item.Year;
 
     var auctionDto = _mapper.Map<AuctionDto>(auction);
-    
+
     await _publishEndpoint.Publish(_mapper.Map<AuctionUpdated>(auctionDto));
 
     var result = await _context.SaveChangesAsync() > 0;
@@ -113,6 +119,8 @@ public class AuctionsController : ControllerBase
     return BadRequest("Problem saving changes");
   }
 
+
+  [Authorize]
   [HttpDelete("{id}")]
   public async Task<ActionResult> DeleteAuction(Guid id)
   {
@@ -123,7 +131,10 @@ public class AuctionsController : ControllerBase
       return NotFound();
     }
 
-    //TODO: check seller == username
+    if(auction.Seller != User.Identity.Name)
+    {
+      return Forbid();
+    }
 
     _context.Auctions.Remove(auction);
